Refactored Song List Detail
Didn't move too many things around because first #32 needs to be
decided.
